feat: ExpositionData 생성 및 모델 구현, refactor:  ExpositionData 파일명 MainExposition으로 수정
